+++ src/video.rs
@@ -21,8 +21,8 @@ impl VideoFrameSpec {
 )]
 #[serde(rename_all = "UPPERCASE")]
 pub enum PixelFormat {
-    Rgb16Be = 0,
-    Rgb16Le = 1,
+    // Rgb16Be = 0,
+    // Rgb16Le = 1,
     #[default]
     Rgb24 = 2,
     Rgb32 = 3,
@@ -31,8 +31,6 @@ pub enum PixelFormat {
 impl PixelFormat {
     pub const fn bytes(self) -> usize {
         match self {
-            PixelFormat::Rgb16Be => 2,
-            PixelFormat::Rgb16Le => 2,
             PixelFormat::Rgb24 => 3,
             PixelFormat::Rgb32 => 4,
         }
@@ -44,8 +42,6 @@ impl PixelFormat {
 
     pub fn from_u8(x: u8) -> Result<Self> {
         match x {
-            0 => Ok(Self::Rgb16Be),
-            1 => Ok(Self::Rgb16Le),
             2 => Ok(Self::Rgb24),
             3 => Ok(Self::Rgb32),
             _ => Err(Failure::new(format!("unknown pixel format: {x}"))),
@@ -78,22 +74,6 @@ impl VideoFrame<Vec<u8>> {
         let d = &mut self.data;
         let i = pos.y as usize * self.spec.stride as usize + pos.x as usize;
         match self.spec.pixel_format {
-            PixelFormat::Rgb16Be => {
-                let r = u16::from(r);
-                let g = u16::from(g);
-                let b = u16::from(b);
-                let v = ((r << 8) & 0xf800) | ((g << 3) & 0x07e0) | (b >> 3);
-                d[i * 2] = (v >> 8) as u8;
-                d[i * 2 + 1] = v as u8;
-            }
-            PixelFormat::Rgb16Le => {
-                let r = u16::from(r);
-                let g = u16::from(g);
-                let b = u16::from(b);
-                let v = ((r << 8) & 0xf800) | ((g << 3) & 0x07e0) | (b >> 3);
-                d[i * 2] = v as u8;
-                d[i * 2 + 1] = (v >> 8) as u8;
-            }
             PixelFormat::Rgb24 => {
                 d[i * 3] = r;
                 d[i * 3 + 1] = g;
@@ -132,22 +112,6 @@ impl<B: AsRef<[u8]>> VideoFrame<B> {
         let d = self.data();
         let i = pos.y as usize * self.spec.stride as usize + pos.x as usize;
         match self.spec.pixel_format {
-            PixelFormat::Rgb16Be => {
-                let v = u16::from(d[i * 2]) << 8 | u16::from(d[i * 2 + 1]);
-                (
-                    (v >> 11) as u8,
-                    (v >> 5) as u8 & 0b111111,
-                    v as u8 & 0b11111,
-                )
-            }
-            PixelFormat::Rgb16Le => {
-                let v = u16::from(d[i * 2]) | u16::from(d[i * 2 + 1]) << 8;
-                (
-                    (v >> 11) as u8,
-                    (v >> 5) as u8 & 0b111111,
-                    v as u8 & 0b11111,
-                )
-            }
             PixelFormat::Rgb24 => (d[i * 3], d[i * 3 + 1], d[i * 3 + 2]),
             PixelFormat::Rgb32 => (d[i * 4], d[i * 4 + 1], d[i * 4 + 2]),
         }